src/backend/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
python-decouple==3.8
psycopg2-binary==2.9.10
Pillow==11.3.0
boto3==1.40.6
django-storages==1.14.6
django-filter==25.1
djoser==2.3.3
PyJWT==2.10.1
redis==6.4.0
celery==5.5.3
gunicorn==23.0.0
whitenoise==6.9.0
python-magic==0.4.27
